+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [06.20.251456]
+
+### Major Changes
+- Working Grading Sheet Page for Faculty.
+- Working Grades Page for Student
+
 
 ## [06.20.250229]